Update find-bootstrap-duration script

+++ scripts/find-bootstrap-duration.sh
@@ -21,8 +21,13 @@ if [[ -n $DATE_OVERRIDE ]]; then
   date="$DATE_OVERRIDE"
 fi
 
+# Determine if the remote script is already in place
+remote_script_exists=$(ssh -o StrictHostKeyChecking=no "$host" 'if [[ -f /tmp/find-bootstrap.sh ]]; then echo "exists"; else echo "not-exists"; fi' 2>/dev/null)
+
 if [[ -n $host ]]; then
-  scp -o StrictHostKeyChecking=no -q "$0" "$host:/tmp/find-bootstrap.sh"
+  if [[ $remote_script_exists == "not-exists" ]]; then
+    scp -o StrictHostKeyChecking=no -q "$0" "$host:/tmp/find-bootstrap.sh"
+  fi
   ssh -o StrictHostKeyChecking=no "$host" "DATE_OVERRIDE=$DATE_OVERRIDE bash /tmp/find-bootstrap.sh"
   exit 0
 fi
